fix: Add remaining tests
